third_party/blink/web_tests/wpt_internal/prerender/session-history-navigation.https.html

<!DOCTYPE html>
<title>Test history.length</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/session-history-test-util.js"></script>
<body>
  <script>
    // We test only a fragment-navigation because other kinds of the main frame navigation
    // in prerender cancels prerendering.
    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testSyntheticAnchorClickInPrerender"),
        "Passed"
      );
    }, "Synthetic anchor click navigates independently with replacement in a prerender");
  </script>
</body>
